+++ Implementation/core/src/main/java/com/teamonehundred/pixelboat/entities/ObstacleFloatingBranch.java
@@ -18,7 +18,7 @@ public class ObstacleFloatingBranch extends ObstacleBranch {
      * @author James Frost
      * @author William Walton
      */
-    public ObstacleFloatingBranch(int x, int y) {
+    public ObstacleFloatingBranch(float x, float y) {
         super(x, y);
         // have the branch move at a constant speed
         drag = 0;
